Create minilua binary with host compiler (fix #80717)
During OpenWrt "cross-compiling" for x86_64, we noticed that during the build
a minilua compiler is generated and finally invoked.

While this is fine when doing a native build, it fails for "cross-compiling"
e.g. due to different used C libraries (e.g. target=musl, host=glibc).

Our solution is to convert the $(CC) to $(HOSTCC) which is then passed via
environment. If not, preset HOSTCC to CC.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>