Fix #80259: Bad handling of RFC 2047 in case of folding white space
While it is correct to strip all whitespace (and even FWS) between two
encoded words[1], it is not correct to strip whitespace between an
encoded-word and text.  Fixing this is trivial; we just have to delay
outputting the space, and can then strip as appropriate.

However, that breaks two tests in the suite, because the current
implementation collapses multiple whitespace after CRLF into a single
space.  This violates RFC 5322 which states[2]:

| Unfolding is accomplished by simply removing any CRLF
| that is immediately followed by WSP.

Changing this may break some code, though, and it's especially doubtful
to do it because `mb_decode_mimeheader()` exhibits the same behavior.

Thus, I'm dropping this PR for further discussion.

[1] <https://tools.ietf.org/html/rfc2047#section-8>
[2] <https://tools.ietf.org/html/rfc5322#section-2.2.3>