Fix #80931: HTTP stream hangs if server doesn't close connection
As is, we read HTTP streams until the server closes the connection,
what is standard behavior prior to HTTP/1.1 anyway.  However, HTTP/1.1
defaults to not closing the connection, and although we are sending a
`Connection: close` header, there is no guarantee that the server heeds
it.  Even worse, if a user passes a `Connection: keep-alive` header via
the stream context, the server is not supposed to close the connection,
so the request hangs (and will only terminated by a timeout).

The proper way to read HTTP responses is to the heed the sent Content-
Length, or in case of `Transfer-encoding: chunked`, to read until after
the final chunk (zero bytes).  To avoid baking this into the general
streams behavior, we use stream filters for this purpose (as suggested
by @IMSoP), and if the end of the stream is detected by those, we set
the `eof` flag.  Since there are already the quite suitable filters for
our purpose (`dechunk` and `consumed`) we attach the desired behavior
to these.

We avoid the filter overhead in cases where the server is going to
close the connection anyway, i.e. prior to HTTP/1.1, or if the server
responds with `Connection: close`.

Currently, the end of content check is enabled for the existing
filters by passing filter parameters.  I kept that simple for now, by
just passing a bool or a long, but that might be to big a BC break;
neither the `dechunk` nor the `consumed` filter are documented, but
easily recognizable via `stream_get_filters()`, so existing code may
already pass parameters, although these are so far unused.  That could
be mitigated by passing an array with respective members instead.
Another option to avoid the BC altogether would be to add the required
info to the stream name (similar to the `convert.iconv.*` filter).

For best BC, it might also be desireable to only set the `eof` flag
from the `dechunk` filter, if `dechunk` is explicitly requested.