Add ASan support for VS
As of VS 16.9, there is native support for ASan in the MSVC toolchain.
Although it is already possible to enable that by adding the respective
option to `CFLAGS`, we make it available via the configure option
`--enable-sanitizer` which is only used for clang builds on Windows so
far.

We also disable the manual inclusion of ASan headers, since these are
available by default.  We use the more general `__SANITIZE_ADDRESS__`
for detection of ASan, to avoid a bogus warning, when `_setmaxstdio()`
is called.