JIT/AArch64: [macos] Workaroud for ZTS
It's not easy for JIT to obtain the address of thread local variable in
Macos as the code emitted by compiler would be patched by linker
afterwards [1].

This patch is one workaround to support ZTS in Macos Apple Silicon.
Runtime function tsrm_get_ls_cache() would be invoked by JIT to obtain
the needed address and all caller-saved registers should be saved/stored
before/after the invocation.

W^X protection is enabled in Macos, i.e. "a thread cannot write to a
memory region and execute instructions in that region at the same time"
[2]. This would affect the behaviors under some jit_debug options and we
will handle it in the near future.

With this patch, all ~4k test cases can pass for JIT/arm64 in ZTS mode
with Macos Apple Silicion.

[1] http://www.jakubkonka.com/2021/01/21/llvm-tls-apple-silicon.html
[2]
https://developer.apple.com/documentation/apple-silicon/porting-just-in-time-compilers-to-apple-silicon

Change-Id: I2e31b97dbeed1504d10665059d5bb4031e98c5c1