JIT/AArch64: Use reserved TMP register for macros rather than REG0/1/2 if possible
REG0/1/2 are passed to the following macros and used as temporary
registers inside. We change to use the reserved TMP register for
AArch64.

```
  EXT_CALL
  JMP_IP
  EXT_JMP
  ADDR_STORE
  SET_EX_OPLINE
  SET_ZVAL_LVAL
  ADDREF_CONST
  MEM_LOAD_64_ZTS
  MEM_LOAD_32_ZTS
  MEM_STORE_64_ZTS
```

REG0 is hard-coded in the following macros and used as temporary
registers as well. We add new argument 'tmp_reg' and use our reserved
TMP register at use sites.

```
  UNDEFINED_OFFSET
  UNDEFINED_INDEX
  CANNOT_ADD_ELEMENT
  EFREE_REFERENCE
  UNDEF_OPLINE_RESULT
  SEPARATE_ARRAY
```

REG0 is hard-coded in macro EMALLOC and used to hold the address of
allocated space. It's better to add new argument 'reg' and pass REG0 at
the use sites.

Besides, several code-style issues are fixed.

Change-Id: I11256b62619e1fa67621e2215ca5ff6f6a0d9177