Fix bug #81096: Re-infer ranges if ref type is inferred
If a reference type for a variable is inferred, the range
information becomes potentially incorrect, as the variable may
be modified indirectly at any time.

If a reference type is inferred, set the variable to the full
range and recompute dependent ranges. This is done using a separate
range worklist that calculates ranges using a widening meet.

Fix for https://bugs.php.net/bug.php?id=81096.