DynASM/arm64: Partially support 'movi' instruction
In the development of PHP JIT/AArch64, `movi d0, #0` is needed to zero
FP registers.

In this patch, we partially support the 64-bit scalar variant of 'movi'
instruction [1].

In this variant, 64-bit immediate value is acceptable with the condition
that each byte is either all zeros or all ones.

Note that the action to handle variables for this 64-bit immediate is
not implemented yet.

Example:
```
  |  movi d0, #0xffff00ffff
  |  movi Rd(reg-ZREG_V0), #0
```

Disassembly:
```
  movi    d0, #0xffff00ffff
  movi    d0, #0x0
```

[1]
https://developer.arm.com/documentation/ddi0602/2020-12/SIMD-FP-Instructions/MOVI--Move-Immediate--vector--?lang=en

Change-Id: I4d3c90e616cbd311437f8d25019ffc60f4090e31