Enable net_get_interfaces on IBM i PASE
On IBM i, getifaddrs is available renamed as Qp2getifaddrs; freeifaddrs
and struct ifaddr follow a similar fate. I assume this is in case AIX
decides to add it (as if). As such, just use the same implementation,
but rename the standard ones to the IBM i definitions and include its
header.

On PHP 7.x, an ifdef to basic_functions.c is also needed, otherwise the
function will not be registered. Technically, the fact PHP 8 always
registers this function is a bug, but hasn't been caught, it seems.