Add a case to version compare
OSX have a string at the end of their version which stops the verson
comparison working.

The bug was ending in the '#N#' case and comparing to an emprty string

If there is nothing left to compare this change prevents that. I have
added the test case in for this.

The closing bracket is what causes the bug in the current PHP version