Fix #81225: Incorrect immediate encoding when using LEA optimization
In JIT/x86, adding with constants would be optimized to LEA instruction.
One premise is that this addition would not overflow. This applys for
subtraction with constants as well.

In the addressing model of x86_64, the "displacement" field in "base +
dispalcement" mode is 32-bit long and would be sign-extended to 64 bits.
However, in JIT/x86, such range check is missing.

In this patch, we add these range checks and one unit test case.

Note: this bug doesn't affect JIT/arm64 in master branch as such "LEA"
optimization is not conducted in JIT/arm64.

Note: this patch should be merged to master branch as well.

[1] https://bugs.php.net/bug.php?id=81225

Change-Id: I6e5cba0958e888d0bcf44db50b1f148094a241f8