Fix #81243: Too much memory is allocated for preg_replace()
If `new_len > alloc_len`, we had a serious issue anyway.  The
comparison only makes some sense if `new_len == alloc_len`, but in that
case we don't need to *re*-allocate.  Only if `new_len == 0` we need to
allocate an empty zend_string.  However, trimming a potentially over-
allocated string appears to be reasonable, so we drop the condition
altogether.