Fix #78374: Warning when closing FTPS connection after FTP_PUT
`ftp_ssl_shutdown()` attempts a bidirectional shutdown handshake, i.e.
we're sending a close_notify alert, and then wait for the peer to send
its close_notify alert.  If the peer doesn't sent it, typically because
it already closed the connection, we should not raise a warning, but
rather ignore this condition.