Break reference after foreach by reference
After `foreach ($array as &$val) {}` perform a reference-unwrap
on `$val`. This means that it will become a normal value rather
than a reference to the last element of `$array`.

This fixes the common WTF where

    foreach ($array as &$val) {}
    foreach ($array as $val) {}

will change the second-to-last element of the array to be the
same as the last element.