Proposal: Allow reserved keywords in more element declaration names
This proposes to allow certain reserved keywords to be allowed as identifiers
in the following places (in PHP 8.2):

- Global function declarations
- Global constant declarations
- Named class declarations
- Trait declarations
- Interface declarations
- Enum declarations

(They were already allowed in class constant declarations, method declarations,
etc.)

For example, if an application was using `Match` as a class/function name
prior to php 8.0, this declaration would have started being a syntax error
in PHP 8.0.
But with this PR, it would be possible to create element names using those
keywords, or instantiate it or use to it as
`\Match` instead of `Match`,
thanks to https://wiki.php.net/rfc/namespaced_names_as_token
having already been merged in php 8.0.

- (otherwise, workarounds such as class_alias would have been required to use
  those identifiers as a class name, for example)

This aims to reduce the backwards compatibility impact of new reserved keyword
additions in new/recent PHP versions.

Using keywords that can already be types (e.g. static, int, float, never)
continues to be a fatal error in class names.

```
// reserved_non_modifiers:
  INCLUDE | INCLUDE_ONCE | EVAL | REQUIRE | REQUIRE_ONCE | OR | XOR | AND
| INSTANCEOF | NEW | CLONE | EXIT | IF | ELSEIF | ELSE | ENDIF | ECHO | DO | WHILE | ENDWHILE
| FOR | ENDFOR | FOREACH | ENDFOREACH | DECLARE | ENDDECLARE | AS | TRY | CATCH | FINALLY
| THROW | USE | INSTEADOF | GLOBAL | VAR | UNSET | ISSET | EMPTY | CONTINUE | GOTO
| FUNCTION | CONST | RETURN | PRINT | YIELD | LIST | SWITCH | ENDSWITCH | CASE | DEFAULT | BREAK
| ARRAY | CALLABLE | EXTENDS | IMPLEMENTS | NAMESPACE | TRAIT | INTERFACE | CLASS
| __CLASS__ | __TRAIT__ | __FUNCTION__ | __METHOD__ | __LINE__ | __FILE__ | __DIR__ | __NAMESPACE__
| FN | MATCH | ENUM

// semi_reserved:
| STATIC | ABSTRACT | FINAL | PRIVATE | PROTECTED | PUBLIC | READONLY
```

Alternatives Considered
-----------------------

1. One possibility was to instead only allow newly introduced keywords
   since some arbitrary php version (e.g. 7.0).

   This was rejected because decades from now, the choice of which keywords
   are forbidden would seem arbitrary and hard to remember.
2. Another possibility would be to allow fully qualified names in these declarations.
   (e.g. `class \Match`).

   This was rejected since allowing code in namespaces to declare elements
   in different namespaces was much more drastic than the approach taken here.