Allocate zend_class_entry on the heap
I believe that after the immutability changes in PHP-8.1, it's no
longer necessary to allocate zend_class_entry (or anything it
references) on the arena, it should be fine to allocate it on the
heap.

My main motivation here is to make it easier to detect class
use-after-frees after persist, but this also seems reasonable to
avoid leaving behind unused memory in that case.