Zend/tests/bug74093.phpt: Fix failing test case
This test case fails (on non-Windows hosts, where it is enabled) due
to mismatching output in the error log language. This fixes the
expectation, and also rewrites the test procedure in a more stable
fashion.

The objective of the test case is to run a program that exceeds
the max_execution_time and verify that the process was aborted. The
previous implementation tested this using a loop on array_intersect with
large enough inputs to "probably" take enough time to trigger
max_execution_time to abort it. With faster CPUs, over time this test
can become flaky. Instead we simply spin a loop until enough
wall clock time has passed to check our assertion.