Fix #81500: Interval serialization regression since 7.3.14 / 7.4.2
While it may not be desired, `DateInterval::$f` supports negative
values, at least with regard to calculations.  We still need to guard
from assigning double values which are out of range for signed 64bit
integers (which would be undefined behavior).  While we could restrict
this only to values which are actually out of range, the marker for
invalid µs is `-1000000`, so we cannot accept smaller numbers.  To
retain symmetry, we also reject values greater than `1000000`.  We must
not do that only for unserialization, but also when the property is set
in the first place.