Use zend_dval_to_lval() to take care of overflow
If overflow would occur, the value is set to `0`, what is standard PHP
behavior anyway.  This way we can avoid setting the invalid marker,
which doesn't work as expected anyway.

We need to adapt some of the existing tests wrt. this behavior.  In
particular, we check for an arbitrary value in bug79015.phpt, to cater
to differences between 32bit and 64bit architectures.