Fix #81543: parse_url omits leading slash in windows paths
According to the RFC8089 Section 2[1]:

> The generic syntax in [RFC3986] includes "path" and "authority"
  components, for each of which only a subset is used in the definition
  of the file URI scheme.  The relevant subset of "path" is "path-
  absolute"

'path-absolute' in RFC3986 is defined as beginning with a '/' character.

There's also a requirement in RFC3986 section 3.3 [2], that

> If a URI contains an authority component, then the path component
  must either be empty or begin with a slash ("/") character.

For file:///c:/ URL the authority component is present but empty,
which implies 'localhost'. Thus the 'path' component must begin
with the '/' character.

For all intents and purposes 'file:///c:/' and 'file://localhost/c:/'
URLs are equivalent, so parsing them differently is a surprising behaviour.

Other popular languages like Java or JavaScript preserve the
leading slash character as well.

This effectively reverts commit [3].

[1] https://datatracker.ietf.org/doc/html/rfc8089#section-2
[2] https://datatracker.ietf.org/doc/html/rfc3986#section-3.3
[3] https://github.com/php/php-src/commit/4505a611ba8cdcc08505a5ffa0e35bda385bd374