Refactor GC data traversal loops
Previusly, to redce stack usage GC traversed object properties and array
elements in both directions. At first, it seeked backward from the last
element, to find the last refcounted zval. Then, it traversed elements
forward. As result, we had 2 non-spatial memory acesses for each
object/array and usuall 2 expensive CPU cache misses.

Now all the traversal algorithms are refactored to scan elements only in
foeward direction. The number of checks and GC stack usage should be the
same. The order of element processing may be differenr, but this
should not be a problem.