Reduce memory usage when unserializing packed arrays of size 1
Previously, PHP's unserialize() would unconditionally convert a packed array to
an associative(hashed) array to solve the problem of `[0 => $v1, 'key' => $v2]`
starting out as a packed array but becoming an associative array,
causing the raw zval pointer to the value for $v1 to have changed.

- This can only happen when there's at least two elements, though.

Additionally, reduce memory usage when calling `__unserialize` with arrays of
size 0 or 1
(e.g. for user-defined functions that store the passed in array as a property)