Try fixing reference counting bugs from bugfix
Hopefully, the original string can't be released in
`zend_assign_to_string_offset` if we're working on a copy instead,
because the zend_string pointer in the zval was immediately replaced.

See discussion on 09547c64c28f81957d986b9c6b995b28d1098db6

That change would leak memory if the original string was a temporary string
which had 2 or more references, e.g.

```php
$data2 = bzcompress($string, 1, 10);

$data3 = $data2;
$data3[3] = 0;
```