Also cache the computed stack size in INIT_FCALL_BY_NAME
The computed stack size only depends on the function definition fbc and the number
of opcodes in extended_value, and that doesn't change.

INIT_FCALL_BY_NAME is used when the function definition is not available when a
file is being compiled, commonly due to the function being declared in a
different file.
Precomputing the stack size is an optimization already done by INIT_FCALL
used when the function is known.

Replacing `hallo` in Zend/micro_bench.php with a require_once to load `hallo`
from a different file in a build with opcache enabled (but not the jit):

```
Before (best run) - total time and subtracting time needed for an empty loop
empty_loop         0.162
func_elsewhere()   0.615    0.475

After (best run):
empty_loop         0.159 (unaffected, just noise)
func_elsewhere()   0.579    0.420
```