Add SORT_STRICT option to array_unique()
Provide a way to request the use of strict comparison (===) when using
`array_unique`.  SORT_STRICT does no type conversion: values of
different types are ordered by their type.  For array/object values
which are == but not === we recursively compare the
elements/properties.  Arrays with identical elements will be identical
themselves; but for objects with identical properties which are not
identical we break the tie using the `spl_object_id()` of the objects.