Don't involve PHP in Apache mpm_winnt control process
Apache mpm_winnt uses a single control process which launches a single
child process which in turn creates threads to handle requests.  Since
the child process is not forked, but rather spawned, and the control
process is never involved in request handling, there is no need to
power up PHP for the control process.  Besides not doing this saves
some resources, we no longer have to deal with potential re-attaching
to OPcache shared memory which avoids issues due to ASLR, and paves the
way to further improvements (such as preloading support for Apache
mpm_winnt).