Fix GH-7910: rename fails on Windows if the target is being executed
Prior to commit c732ab4[1], the script file was closed immediately
after compilation, what made it possible to replace that file with
another one on Windows (e.g. for hot patching).  This is no longer
possible, since even opening the file with `FILE_SHARE_DELETE`
semantics does not allow to have it as target of `rename()`.

Thus, we revert this part of the `zend_string` refactoring of the Zend
stream API.

[1] <https://github.com/php/php-src/commit/c732ab400af92c54eee47c487a56009f1d79dd5d>