ext/standard/var: optimize php_append_string_export()
The old algorithm was rather clumsy about escaping NUL bytes; for
example:

 var_export("foo\0bar") -> 'foo' . "\0" . 'bar'

These two concatentations result in a degenerate AST with three
levels, causing zend_eval_const_expr() to be called recursively three
times.  It would be better to use double quotes for the whole string,
for example:

 var_export("foo\0bar") -> "foo\0bar"

More NUL bytes cause more and more concatenations, for example:

 var_export("\0\0") -> '' . "\0" . '' . "\0" . ''

Two NUL bytes already cause a zend_eval_const_expr() recursion depth
of 5 (because the AST is so degenerate, it is effectively only a
linked list).

Now try this one:

 var_export(str_repeat("\0", 100000))

Attempting to parse its output with PHP leads to a stack overflow with,
because each zend_eval_const_expr() allocates 128 bytes on the stack
(on x86_64), crashing PHP with SIGSEGV.

This is not a theoretical problem - we observed PHP crashes in a Typo3
installation when Typo3 parsed its PackageManager cache file, which
contains a serialized complex PHP object with lots of protected
variables (that's were NUL bytes come from), and this serialized
string is then passed to var_export().

This is the Typo3 code location where var_export() gets called with an
embedded serialize() call:

 https://github.com/TYPO3/typo3/blob/main/typo3/sysext/core/Classes/Package/Cache/PackageCacheEntry.php#L135

This is similar to bug report https://bugs.php.net/bug.php?id=80183
which was found by a fuzzer.

My solution is to use single quotes only if there are no NUL bytes,
and if there is at least one NUL byte, use a single double quoted
string (without any concatenations).