[ci skip] [Zend] Update README.md with new Zend VM executors
Closes GH-8008

As a new learner to PHP Zend VM, the first place I can find information
is this README.md. Unfortunately, it has not included the latest Zend VM
kind and code generation script.

With code reading and trail-and-error experiments, I found a few updates
that might be helpful to new learners.

ChangeLog:
1) Add HYBRID threading model
2) Remove --without-old-executor option and description
3) ZE2 --> ZE3 (Since PHP 7)
4) Change default VM kind from CALL to HYBRID

Signed-off-by: Su, Tao <tao.su@intel.com>